+++ etils/__init__.py
@@ -16,7 +16,7 @@
 
 # A new PyPI release will be pushed everytime `__version__` is increased
 # When changing this, also update the CHANGELOG.md
-__version__ = '0.5.1'
+__version__ = '0.6.0'
 
 # Do NOT add anything to this file. This is left empty on purpose.
 # Users should import subprojects directly.